requirements.txt
# Core Data Science Libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Dataset Loading
kagglehub
kaggle

# Deep Learning & Graph Neural Networks
torch==2.0.1
torchvision==0.15.2
dgl==1.1.2
torch-geometric==2.3.1

# Graph Database
neo4j==5.12.0
py2neo==2021.2.4

# API Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
python-multipart==0.0.6

# LLM & AI Agent Framework
langchain==0.0.284
langchain-community==0.0.5
google-generativeai==0.2.2
openai==0.28.1

# MLOps & Experiment Tracking
mlflow==2.7.1
dvc==3.27.0
dvc[s3]==3.27.0

# Environment & Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.24.1

# Data Visualization (for notebooks)
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Utilities
tqdm==4.66.1
joblib==1.3.2
networkx==3.1

# Development Tools
black==23.7.0
flake8==6.0.0
isort==5.12.0
mypy==1.5.1

# Jupyter Notebook Support
jupyter==1.0.0
ipykernel==6.25.2

